Develop/test/#38

+++ backend/src/main/java/com/todayyum/member/application/RemoveFollowUseCase.java
@@ -1,5 +1,7 @@
 package com.todayyum.member.application;
 
+import com.todayyum.global.dto.response.ResponseCode;
+import com.todayyum.global.exception.CustomException;
 import com.todayyum.member.application.repository.FollowRepository;
 import com.todayyum.member.domain.Follow;
 import lombok.RequiredArgsConstructor;
@@ -19,6 +21,10 @@ public class RemoveFollowUseCase {
     @Transactional
     public void removeFollow(UUID fromMemberId, UUID toMemberId) {
         Follow follow = Follow.createFollow(fromMemberId, toMemberId);
+
+        if(!followRepository.existsByFromMemberAndToMember(follow)) {
+            throw new CustomException(ResponseCode.NOT_FOLLOW);
+        }
 
         followRepository.deleteByFromMemberAndToMember(follow);
     }
